PRODIGY_BD_05/requirements.txt
Flask
flask-sqlalchemy
flask-jwt-extended
psycopg2-binary
python-dotenv
marshmallow